Splited models.py file to models.py files in particular (blueprint) folders
