+++ app/postShop/views.py
@@ -1,7 +1,8 @@
 from .. import db
 from flask import render_template, request, url_for, redirect, flash, Blueprint
 from flask_login import login_required, current_user
-from ..models import User, Posts
+from .models import Posts
+from ..userShop.models import User
 from ..userShop.views import current_role, role_req
 import time
 import logging
