+++ app/userShop/forms.py
@@ -1,7 +1,7 @@
 from flask_wtf import FlaskForm
 from wtforms import StringField, PasswordField, SubmitField
 from wtforms.validators import DataRequired, Email, ValidationError
-from app.models import User
+from .models import User
 
 
 class RegisterForm(FlaskForm):
