+++ app/userShop/models.py
@@ -1,22 +1,6 @@
 from flask_login import UserMixin
 from werkzeug.security import check_password_hash, generate_password_hash
-
 from app import db, login
-
-
-class Product(db.Model):
-    __tablename__ = 'product'
-
-    id = db.Column(db.Integer, primary_key=True)
-    name = db.Column(db.String(), nullable=False)
-    price = db.Column(db.Float(), nullable=False)
-    quantity = db.Column(db.Integer, nullable=False)
-    description = db.Column(db.String())
-    image = db.Column(db.String())
-
-    def __repr__(self):
-
-        return f'{self.name} {self.price} {self.id}'
 
 
 class User(UserMixin, db.Model):
@@ -57,27 +41,6 @@ class UserRoles(db.Model):
     id = db.Column(db.Integer(), primary_key=True)
     user_id = db.Column(db.Integer(), db.ForeignKey('user.id', ondelete='CASCADE'))
     role_id = db.Column(db.Integer(), db.ForeignKey('role.id', ondelete='CASCADE'))
-
-
-class Posts(db.Model):
-    __tablename__ = 'posts'
-    id = db.Column(db.Integer(), primary_key=True)
-    title = db.Column(db.String(100), unique=True, nullable=False)
-    post = db.Column(db.String(), nullable=False)
-    user = db.Column(db.String(), nullable=False)
-    time = db.Column(db.Float())
-    time_date = db.Column(db.String())
-
-    def __repr__(self):
-        return f'{self.title} {self.post}'
-
-
-class Cart(db.Model):
-    __tablename__ = 'cart'
-    id = db.Column(db.Integer(), primary_key=True)
-    user_id = db.Column(db.Integer(), db.ForeignKey('user.id', ondelete='CASCADE'))
-    product_id = db.Column(db.Integer(), db.ForeignKey('product.id', ondelete='CASCADE'))
-    quantity = db.Column(db.Integer(), nullable=True)
 
 
 @login.user_loader
